requirements.txt
duckdb
pandas
pyarrow
boto3
jproperties
google-cloud-bigquery
google-cloud-storage
